requirements.txt
Flask
bcrypt
Pillow
Flask-Session
